fix(models): repair grouped picker rendering; clarify sheet purpose
- The dynamic picker accidentally nested Sections inside a Section,
  which List renders incorrectly: provider headers swallowed model
  rows (rows showed no model names) and tap hit-testing broke. The
  picker now renders as sibling sections via a shared
  ModelPickerSection view (search field + one section per connected
  provider, flat rows).
- Model show/hide stays entirely server-side: the picker renders the
  /provider registry's connected set (server-configured providers +
  keyless local ones), no client-side model configuration UI.
- Rename the sheet title Configure -> Model & Agent so its runtime
  switching role is obvious.
