+++ OpenCodeClient/OpenCodeClient/Views/Chat/ChatToolbarView.swift
@@ -132,7 +132,13 @@ struct ChatToolbarView: View {
         .sheet(isPresented: $showConfigSheet) {
             NavigationStack {
                 List {
-                    modelPickerSection
+                    if !state.dynamicModelPresets.isEmpty {
+                        Section {
+                            TextField(L10n.t(.configureModelSearchPlaceholder), text: $modelSearchText)
+                                .autocorrectionDisabled()
+                        }
+                    }
+                    ModelPickerSection(state: state, searchText: $modelSearchText)
 
                     Section(L10n.t(.configureAgent)) {
                         if state.isLoadingAgents {
@@ -187,84 +193,7 @@ struct ChatToolbarView: View {
 
     // MARK: - Model Picker
 
-    /// Groups the dynamic picker list by provider, applying the search
-    /// filter. Entries keep their flat indices (via pickerModelPresets) so
-    /// selection bookkeeping stays index-based.
-    private var modelPickerSection: some View {
-        let presets = state.pickerModelPresets
-        let query = modelSearchText.trimmingCharacters(in: .whitespacesAndNewlines).lowercased()
-        let filtered = query.isEmpty
-            ? presets
-            : presets.filter {
-                $0.displayName.lowercased().contains(query)
-                    || $0.modelID.lowercased().contains(query)
-                    || $0.providerID.lowercased().contains(query)
-            }
-        let grouped: [(providerID: String, entries: [(index: Int, preset: ModelPreset)])] = {
-            var byProvider: [String: [(Int, ModelPreset)]] = [:]
-            var order: [String] = []
-            for (idx, preset) in filtered.enumerated() {
-                let pid = preset.providerID
-                if byProvider[pid] == nil { order.append(pid) }
-                byProvider[pid]?.append((idx, preset))
-            }
-            return order.map { ($0, byProvider[$0] ?? []) }
-        }()
-
-        return Section {
-            if !state.dynamicModelPresets.isEmpty {
-                // Dynamic mode: searchable, grouped by connected provider.
-                Section {
-                    TextField(L10n.t(.configureModelSearchPlaceholder), text: $modelSearchText)
-                        .autocorrectionDisabled()
-                }
-                ForEach(grouped, id: \.providerID) { group in
-                    Section(
-                        header: Text(state.providerDisplayNames[group.providerID] ?? group.providerID)
-                    ) {
-                        ForEach(group.entries, id: \.preset.id) { entry in
-                            modelRow(index: entry.index, preset: entry.preset)
-                        }
-                    }
-                }
-                if filtered.isEmpty {
-                    Text(L10n.t(.configureModelNoMatches))
-                        .foregroundColor(.secondary)
-                }
-            } else {
-                // Fallback mode: curated presets (registry unavailable).
-                Section(L10n.t(.configureModel)) {
-                    ForEach(Array(presets.enumerated()), id: \.element.id) { index, preset in
-                        modelRow(index: index, preset: preset)
-                    }
-                }
-            }
-        }
-    }
-
-    private func modelRow(index: Int, preset: ModelPreset) -> some View {
-        Button {
-            state.setSelectedModelIndex(index)
-        } label: {
-            HStack {
-                VStack(alignment: .leading, spacing: 2) {
-                    Text(preset.displayName)
-                    if !state.dynamicModelPresets.isEmpty && preset.displayName != preset.modelID {
-                        Text(preset.modelID)
-                            .font(.caption2)
-                            .foregroundColor(.secondary)
-                            .lineLimit(1)
-                    }
-                }
-                Spacer()
-                if state.selectedModelIndex == index {
-                    Image(systemName: "checkmark")
-                        .foregroundColor(DesignColors.Brand.primary)
-                }
-            }
-        }
-        .foregroundColor(.primary)
-    }
+    // Model rows live in ModelPickerSection (shared with Settings).
 
     // MARK: - Todo Button & Panel
